requirements.txt
Pillow>=9.0.0
exifread>=2.3.2